Fix stepsize issue and stop sending PBT commands when not supported by mode.

+++ spectrumscope.h
@@ -79,7 +79,7 @@ public:
     void setIFShift (uchar val);
 
     quint16 getStepSize () { return stepSize;}
-    void setStepSize (quint16 hz) { stepSize = hz;}
+    void setStepSize (quint64 hz) { stepSize = hz;}
 
     freqt getFrequency () { return freq;}
     void setFrequency (freqt f,uchar vfo=0);
@@ -279,7 +279,7 @@ private:
     double passbandCenterFrequency = 0.0;
     double pbtDefault = 0.0;
     quint16 cwPitch = 600;
-    quint16 stepSize = 100;
+    quint64 stepSize = 100;
     uchar ifShift = 0;
     int refLower=0;
     int refUpper=0;
